application/requirements.txt
# ---- Web & API ----
Django==4.2.13
djangorestframework==3.14.0
drf-yasg==1.21.7
django-environ== 0.11.0
dj-database-url==2.2.0
django-cors-headers==4.7.0

# ---- DB & Storage ----
psycopg2-binary==2.9.9
django-storages==1.14.4
boto3==1.34.162

# ---- Queue / Async ----
celery==5.4.0
redis==5.0.6

# ---- WebSockets / ASGI ----
channels==4.0.0
channels-redis==4.1.0
uvicorn==0.30.6
gunicorn==22.0.0

# ---- CV / Imaging ----
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
scikit-image==0.22.0
imutils==0.5.4
pycocotools==2.0.7
joblib==1.4.2

# ---- DL (CPU wheels available on PyPI) ----
torch==2.3.1
torchvision==0.18.1

# ---- Re-ID indexing ----
annoy==1.17.3

# (drf-yasg dep, listed explicitly to avoid surprises)
uritemplate==4.1.1